OutputManagement fractional scaling
Summary:
In order to have fractional scaling in kwin, we need to communicate it with
kscreen, which means changing the data type in our config protocols.

This introduces a new method on outputdevice and outputconfiguration
to set/request scale as a float.

wl_output is and should remain unchanged as an int

No urgent rush for reviewing/merging this as it's useless without other changes.

Test Plan: Attached unit tests

Reviewers: #kwin, romangg

Subscribers: romangg, zzag, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D13601

+++ src/wayland/server/outputchangeset_p.h
@@ -40,7 +40,7 @@ namespace KWayland
             int modeId;
             OutputDeviceInterface::Transform transform;
             QPoint position;
-            int scale;
+            qreal scale;
         };
     }
 }
